requirements.txt
aiohttp==3.9.5
python-dotenv==0.20.0